+++ src/monocle-ts/lensesFromInterface.ts
@@ -5,7 +5,7 @@ export type LensesFromProps<P extends t.Props, T extends t.InterfaceOf<P> = t.In
   [K in keyof P]: Lens<T, T[K]>
 }
 
-export function lensesFromProps<P extends t.Props>(i: t.InterfaceType<P>): LensesFromProps<P> {
+export function lensesFromInterface<P extends t.Props>(i: t.InterfaceType<P>): LensesFromProps<P> {
   const r: any = {}
   for (const k in i.props) {
     r[k] = Lens.fromProp<any, typeof k>(k)