+++ CHANGELOG.md
@@ -12,6 +12,13 @@
 **Note**: Gaps between patch versions are faulty/broken releases.
 **Note**: A feature tagged as Experimental is in a high state of flux, you're at risk of it changing without notice.
 
-0.0.1
+# 0.0.2
+
+- **New Feature**
+  - add `lensesFromInterface` (@leemhenson)
+  - add `lensesFromTuple` (@gcanti)
+  - add `prismsFromUnion` (@gcanti)
+
+# 0.0.1
 
 Initial release